+++ ts/input/tex/begingroup/BegingroupMethods.ts
@@ -25,7 +25,7 @@ import { CommandMap } from '../TokenMap.js';
 import TexParser from '../TexParser.js';
 import TexError from '../TexError.js';
 
-import { COMPONENT } from './locales/Component.js';
+import { COMPONENT } from './__locales__/Component.js';
 import BaseMethods from '../base/BaseMethods.js';
 import { begingroupStack } from './BegingroupStack.js';
