+++ ts/input/tex/braket/BraketConfiguration.ts
@@ -25,7 +25,7 @@ import { HandlerType, ConfigurationType } from '../HandlerTypes.js';
 import { Configuration } from '../Configuration.js';
 import { BraketItem } from './BraketItems.js';
 import './BraketMappings.js';
-export { COMPONENT } from './locales/Component.js';
+export { COMPONENT } from './__locales__/Component.js';
 
 export const BraketConfiguration = Configuration.create('braket', {
   [ConfigurationType.HANDLER]: {
