+++ ts/input/tex/extpfeil/ExtpfeilConfiguration.ts
@@ -31,7 +31,7 @@ import { AmsMethods } from '../ams/AmsMethods.js';
 import { NewcommandUtil } from '../newcommand/NewcommandUtil.js';
 import { NewcommandConfig } from '../newcommand/NewcommandConfiguration.js';
 import TexError from '../TexError.js';
-import { COMPONENT } from './locales/Component.js';
+import { COMPONENT } from './__locales__/Component.js';
 export { COMPONENT };
 
 // Namespace
