+++ ts/input/tex/mathtools/MathtoolsConfiguration.ts
@@ -42,7 +42,7 @@ import {
 } from './MathtoolsMethods.js';
 import { MathtoolsTagFormat } from './MathtoolsTags.js';
 import { MultlinedItem } from './MathtoolsItems.js';
-export { COMPONENT } from './locales/Component.js';
+export { COMPONENT } from './__locales__/Component.js';
 
 /**
  * Add any pre-defined paired delimiters, and subclass the configured tag format.
