+++ ts/input/tex/newcommand/NewcommandMethods.ts
@@ -33,7 +33,7 @@ import { UnitUtil } from '../UnitUtil.js';
 import { StackItem } from '../StackItem.js';
 import { NewcommandUtil } from './NewcommandUtil.js';
 
-import { COMPONENT } from './locales/Component.js';
+import { COMPONENT } from './__locales__/Component.js';
 
 // Namespace
 const NewcommandMethods: { [key: string]: ParseMethod } = {
