+++ ts/input/tex/setoptions/SetOptionsConfiguration.ts
@@ -36,7 +36,7 @@ import { Macro } from '../Token.js';
 import BaseMethods from '../base/BaseMethods.js';
 import { expandable, isObject } from '../../../util/Options.js';
 import { PrioritizedList } from '../../../util/PrioritizedList.js';
-import { COMPONENT } from './locales/Component.js';
+import { COMPONENT } from './__locales__/Component.js';
 export { COMPONENT };
 
 export const SetOptionsUtil = {
